BAH-2506 | fix. Lab report upload fails when redirected to lab entry from patient dashboard
